meta_details file renamed
