+++ src/state_types/models/meta_details.rs
